Move the services into their own folders. Adjust namespaces.
