+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Functions/CommandHandlers/FullReset/FullResetFunction.cs
@@ -1,5 +1,6 @@
 ﻿using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Functions.CommandHandlers.RefreshSearchableDocument.Dto;
 using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
+using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.ResetSearchableDocuments;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.Functions.Worker;
