+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/ResetSearchableDocuments/FullSearchableDocumentResetter.cs
@@ -1,7 +1,8 @@
 ﻿using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Clients.ContentApi;
+using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.RemoveSearchableDocument;
 using Microsoft.Extensions.Logging;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
+namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.ResetSearchableDocuments;
 
 public class FullSearchableDocumentResetter(
     IContentApiClient contentApiClient,
