+++ src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Services/ResetSearchableDocuments/PerformResetResponse.cs
@@ -1,6 +1,6 @@
 ﻿using GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Domain;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services;
+namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Services.ResetSearchableDocuments;
 
 public class PerformResetResponse
 {
